+++ HISTORY.rst
@@ -7,6 +7,7 @@ Changelog
 * Field-level validation.
 * Add ``fields.Method``.
 * Allow binding of extra data to a serialized object by passing the ``extra`` param when initializing a ``Serializer``.
+* Add ``relative`` paramater to ``fields.Url`` that allows for relative URLs.
 
 0.1.0 (2013-11-10)
 ++++++++++++++++++